requirements.txt
requests
beautifulsoup4
lxml
python-whois
dnspython
selenium
PyYAML
